+++ infra/common/src/main/java/org/apache/shardingsphere/infra/version/ShardingSphereVersion.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.autogen.version;
+package org.apache.shardingsphere.infra.version;
 
 import com.google.common.base.Strings;
 import lombok.AccessLevel;
